backend/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
ecdsa==0.18.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
install==1.3.5
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
PyMySQL==1.0.2
pypng==0.20220715.0
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
qrcode==7.4.2
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
